Add helper for creating session and model
