+++ tests/conftest.py
@@ -5,8 +5,6 @@ import pytest
 import sqlalchemy
 from sqlalchemy import orm
 
-from open_alchemy import models
-
 
 @pytest.fixture(scope="function", params=["sqlite:///:memory:"])
 def engine(request):
@@ -18,23 +16,3 @@ def engine(request):
 def sessionmaker(engine):
     """Creates a sqlite session."""
     return orm.sessionmaker(bind=engine)
-
-
-@pytest.fixture(autouse=True)
-def cleanup_models():
-    """Remove any new attributes on open_alchemy.models."""
-    for key in set(models.__dict__.keys()):
-        if key.startswith("__"):
-            continue
-        if key.endswith("__"):
-            continue
-        delattr(models, key)
-
-    yield
-
-    for key in set(models.__dict__.keys()):
-        if key.startswith("__"):
-            continue
-        if key.endswith("__"):
-            continue
-        delattr(models, key)
